backend/requirements.txt
django == 3.2.12
djangorestframework == 3.14.0
djangorestframework_simplejwt
django-cors-headers == 3.14.0
pymysql